+++ server/errors/BadRequestError.ts
@@ -3,7 +3,7 @@ import { InternalError } from "./InternalError";
 
 export class BadRequestError extends InternalError {
 	constructor(message: string = "Invalid data") {
-		super({ status: HttpStatusCode.BAD_REQUEST, message, harmless: true });
+		super({ status: HttpStatusCode.BAD_REQUEST, code: "unknown", message, harmless: true });
 		this.name = "BadRequestError";
 	}
 }