+++ server/errors/NotEnoughRoomError.ts
@@ -3,9 +3,9 @@ import { InternalError } from "./InternalError";
 
 export class NotEnoughRoomError extends InternalError {
 	constructor(
-		message: string = "There is not enough room to write your data. Delete some stuff first"
+		message: string = "There is not enough room to write your data. Delete some stuff first."
 	) {
-		super({ status: HttpStatusCode.INSUFFICIENT_STORAGE, message });
+		super({ status: HttpStatusCode.INSUFFICIENT_STORAGE, code: "storage-quota-exceeded", message });
 		this.name = "NotEnoughRoomError";
 	}
 }